Removing header files for now
