+++ bloss1/src/ecs/systems/physics_system.cpp
@@ -1,11 +1,22 @@
 #include "ecs/systems.hpp"
-#include "ecs/systems/physics_system.hpp"
 
 #define GRAVITY 9.8f
 
 // @TODO: use continuous collision detection
 namespace bls
 {
+    struct Collision
+    {
+        vec3 point_a;
+        vec3 point_b;
+        bool has_collision;
+    };
+
+    void resolve_collisions(ECS& ecs);
+    Collision test_collision(ECS& ecs, u32 id_a, u32 id_b);
+    void solve_collision(ECS& ecs, u32 id_a, u32 id_b, Collision collision);
+    f32 apply_deceleration(f32 velocity, f32 deceleration, f32 mass, f32 dt);
+
     void physics_system(ECS& ecs, f32 dt)
     {
         resolve_collisions(ecs);
